CrashCourse/requirements.txt
langchain==0.1.3
openai==1.9.0
python-dotenv==1.0.1
langchain-openai==0.0.3
huggingface_hub==0.20.3
streamlit==1.30.0
python-dotenv